+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - Minor cleanup to the `DJISerial` class (including tests).
 - Add some utilities to the `DJISerial::SerialMessage` class.
 - Add generic `RemoteMapState` constructor (!137).
+- Fixed `Profiler` class.
 
 ## January 2022